Simplify instantiation of anonymous classes implementing interfaces.
I'm not sure this is the best way of doing things, but I think it's
better than it was. The "context" instance variable in the anonymous
class is a bit awkward; perhaps we could do something clever with
bindings here? Also in most cases there's probably no need to actually
include the module representing the Java interface, but for the moment I
think it makes it clearer what's going on.