Use JRuby closure conversion instead of my clumsy workaround.
"The block is converted to a Proc object, which is then decorated with a
java interface proxy that invokes the block for *any* method called on the
interface. This not only works for event listeners or Runnable, but
basically for any interface. When calling a method that expects an
interface, JRuby checks if a block is passed and automatically converts
the block to an object implementing the interface." [1]

[1] https://github.com/jruby/jruby/wiki/CallingJavaFromJRuby