Trap unexpected invocation in "Sniper Has More To Say" in chapter 14.
Shows unit test failing with unexpected invocation :-

      1) Failure:
    test_0003_reports_is_winning_when_current_price_comes_from_sniper(AuctionsniperSpec)
[app/auction_sniper.rb:9]:
    unexpected invocation: #<Mock:Auction>.bid(168)
    unsatisfied expectations:
    - expected at least once, not yet invoked:
      #<Mock:SniperListener>.sniper_winning(any_parameters)

So far I haven't been attempting to replicate the existence of
interfaces in the Java-based code from the book. This is the first time
I've felt a bit of pain from this - I decided to introduce PriceSource
as a top-level class, because it's needed by both AuctionSniper and
AuctionMessageTranslator and I have no AuctionEventListener interface in
which to put it.