Failing end-to-end tests in "Displaying Price Details" in chapter 15.
The tests are failing because the table row has no details, only status
text :-

    NativeException: java.lang.AssertionError:
    Tried to look for...
        exactly 1 JTable (ANYTHING)
        in exactly 1 JFrame (with name "Auction Sniper Main" and showing
on screen)
        in all top level windows
    and check that it is table with row with cells <label with text
"item-54321">, <label with text "1000">, <label with text "1098">,
<label with text "bidding">
    but...
        all top level windows
        contained 1 JFrame (with name "Auction Sniper Main" and showing
on screen)
        contained 1 JTable (ANYTHING)
       it is not table with row with cells <label with text
"item-54321">, <label with text "1000">, <label with text "1098">,
<label with text "bidding">
        because
          in row 0: component 0 text was "bidding"

It is unclear, but in the book it looks like
`AuctionSniperDriver#shows_sniper_status` may have been overloaded so
there were different implementations for different sets of parameters.
I've decided to introduce `AuctionSniperDriver#driver.shows_sniper_status_text`
to avoid confusion and to reduce the number of changes needed in this
commit.